Simplify names of diagonal drawing commands and reorganize
Make it so the "Corner" commands used for "X" are just the default, and
specifically label the ones with the flat horizontal or vertical edges.
Also remove the "Segment" from each name for simplicity, and change the
order that the methods are listed. For example:

 o DrawLeftDiagonalSegmentCorner -> DrawLeftDiagonal
 o DrawLeftDiagonalSegment       -> DrawLeftDiagonalHorizontal

This is largely for personal preference, and seemed to make sense before
digging into the diagonal drawing commands to fix the thickness relative
to the vertical and horizontal segments.