Remove "match"-based letter drawing
The match statement was causing problems in PsychoPy because it uses
Python 3.8, and match was introduced in 3.10. I changed that section of
code to if-then-elif in the hopes it would let PsychoPy run this, but it
turns out there are other issues, probably having to do with library
versions.

I plan to revert this change in the next commit, but wanted to keep the
patch around since I did the work. :)